Fix handling of terminal resizing.
